Music and animation generator & getter
-Added music that loops :musical_note:

-Added rawTexturesArray making it possible to add the preffix "raw" to a
file so it isn't automatically assigned a TextureRegion.

-Added dynamic Animation loader that automagically detects the number of
frames in an animation of single row png animation file.
Usage is: getAnimation(float time between frames, file name)
IMPORTANT: The animation's png MUST be of multiple columns, single row
and MUST have raw_ preffix it won't work. Also the animation frame size
must be of equal width and height, e.g. Frame is 40x40 in a 240x40 file.
It's sort of restrictive but it works like a charm.
-Added random funny comments.

I hope you read this, pold10, before you resume your coding, so we avoid
the problems with merging.